+++ temba/contacts/tests.py
@@ -263,9 +263,7 @@ class ContactCRUDLTest(CRUDLTestMixin, TembaTest):
 
         # we re-run the search for the response, but exclude Joe
         self.assertEqual(
-            call(
-                self.org.id, group_uuid=str(active_contacts.uuid), query="Joe", sort="", offset=0, exclude_ids=[joe.id]
-            ),
+            call(self.org.id, group_id=active_contacts.id, query="Joe", sort="", offset=0, exclude_ids=[joe.id]),
             mr_mocks.calls["contact_search"][-1],
         )
 
@@ -2076,7 +2074,8 @@ class ContactTest(TembaTest, CRUDLTestMixin):
 
         with AnonymousOrg(self.org):
             mock_search_contacts.side_effect = [
-                SearchResults(query="", total=1, contact_ids=[self.billy.id], metadata=QueryMetadata())
+                SearchResults(query="", total=1, contact_ids=[self.billy.id], metadata=QueryMetadata()),
+                SearchResults(query="", total=1, contact_ids=[self.billy.id], metadata=QueryMetadata()),
             ]
             self.assertEqual(
                 [
@@ -2089,6 +2088,14 @@ class ContactTest(TembaTest, CRUDLTestMixin):
                     {"id": str(self.billy.uuid), "name": "Billy Nophone", "type": "contact"},
                 ],
                 omnibox_request(""),
+            )
+
+            self.assertEqual(
+                [
+                    # 1 contact
+                    {"id": str(self.billy.uuid), "name": "Billy Nophone", "type": "contact"},
+                ],
+                omnibox_request("search=Billy"),
             )
 
         # exclude blocked and stopped contacts
@@ -2478,7 +2485,7 @@ class ContactTest(TembaTest, CRUDLTestMixin):
 
         # errored show retrying icon
         msg.status = Msg.STATUS_ERRORED
-        self.assertIn('"refresh-cw"', msg_status_badge(msg))
+        self.assertIn('"icon.retry"', msg_status_badge(msg))
 
         # failed messages show an x
         msg.status = Msg.STATUS_FAILED
@@ -2562,41 +2569,37 @@ class ContactTest(TembaTest, CRUDLTestMixin):
         )
 
     def test_get_scheduled_messages(self):
-        self.just_joe = self.create_group("Just Joe", [self.joe])
+        just_joe = self.create_group("Just Joe", [self.joe])
 
-        self.assertFalse(self.joe.get_scheduled_broadcasts())
+        self.assertEqual(0, self.joe.get_scheduled_broadcasts().count())
 
         broadcast = Broadcast.create(self.org, self.admin, {"eng": "Hello"}, contacts=[self.frank])
-        self.assertFalse(self.joe.get_scheduled_broadcasts())
+        self.assertEqual(0, self.joe.get_scheduled_broadcasts().count())
 
         broadcast.contacts.add(self.joe)
 
-        self.assertFalse(self.joe.get_scheduled_broadcasts())
+        self.assertEqual(0, self.joe.get_scheduled_broadcasts().count())
 
         schedule_time = timezone.now() + timedelta(days=2)
         broadcast.schedule = Schedule.create_schedule(self.org, self.admin, schedule_time, Schedule.REPEAT_NEVER)
-        broadcast.save()
+        broadcast.save(update_fields=("schedule",))
 
         self.assertEqual(self.joe.get_scheduled_broadcasts().count(), 1)
-        self.assertTrue(broadcast in self.joe.get_scheduled_broadcasts())
+        self.assertIn(broadcast, self.joe.get_scheduled_broadcasts())
 
         broadcast.contacts.remove(self.joe)
-        self.assertFalse(self.joe.get_scheduled_broadcasts())
+        self.assertEqual(0, self.joe.get_scheduled_broadcasts().count())
 
-        broadcast.groups.add(self.just_joe)
+        broadcast.groups.add(just_joe)
         self.assertEqual(self.joe.get_scheduled_broadcasts().count(), 1)
-        self.assertTrue(broadcast in self.joe.get_scheduled_broadcasts())
+        self.assertIn(broadcast, self.joe.get_scheduled_broadcasts())
 
-        broadcast.groups.remove(self.just_joe)
-        self.assertFalse(self.joe.get_scheduled_broadcasts())
-
-        broadcast.urns.add(self.joe.get_urn())
-        self.assertEqual(self.joe.get_scheduled_broadcasts().count(), 1)
-        self.assertTrue(broadcast in self.joe.get_scheduled_broadcasts())
+        broadcast.groups.remove(just_joe)
+        self.assertEqual(0, self.joe.get_scheduled_broadcasts().count())
 
         broadcast.schedule.next_fire = None
         broadcast.schedule.save(update_fields=["next_fire"])
-        self.assertFalse(self.joe.get_scheduled_broadcasts())
+        self.assertEqual(0, self.joe.get_scheduled_broadcasts().count())
 
     def test_update_urns_field(self):
         update_url = reverse("contacts.contact_update", args=[self.joe.pk])