+++ temba/msgs/views.py
@@ -148,6 +148,10 @@ class MsgListView(SpaMixin, ContentMenuMixin, OrgPermsMixin, BulkActionMixin, Sm
 
     def build_content_menu(self, menu):
         if self.is_spa():
+            if self.has_org_perm("msgs.broadcast_send"):
+                menu.add_modax(
+                    _("Send Message"), "send-message", reverse("msgs.broadcast_send"), title=_("Send Message")
+                )
             if self.has_org_perm("msgs.label_create"):
                 menu.add_modax(_("New Label"), "new-msg-label", reverse("msgs.label_create"), title=_("New Label"))
 
@@ -219,7 +223,7 @@ class BroadcastCRUDL(SmartCRUDL):
                 .get_queryset(**kwargs)
                 .filter(is_active=True)
                 .select_related("org", "schedule")
-                .prefetch_related("groups", "contacts", "urns")
+                .prefetch_related("groups", "contacts")
             )
 
     class ScheduledCreate(OrgPermsMixin, ModalMixin, SmartFormView):
@@ -748,7 +752,7 @@ class MsgCRUDL(SmartCRUDL):
                     org=self.request.org, status=Broadcast.STATUS_QUEUED, schedule=None, is_active=True
                 )
                 .select_related("org")
-                .prefetch_related("groups", "contacts", "urns")
+                .prefetch_related("groups", "contacts")
                 .order_by("-created_on")
             )
             return context